+++ src/launch.js
@@ -15,11 +15,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { MainRenderDriver, RendererMain, ThreadXRenderDriver } from '@lightningjs/renderer'
+import { MainCoreDriver, RendererMain } from '@lightningjs/renderer'
 // import RendererWorker from '@lightningjs/renderer/workers/renderer?worker'
 import Settings from './settings.js'
 import { initLog, Log } from './lib/log.js'
-
+import { screenResolutions } from './lib/utils.js'
 // import coreExtensionModule from './fontLoader.js?importChunkUrl'
 
 export let renderer
@@ -29,18 +29,23 @@ export default (App, target, settings) => {
 
   initLog()
 
-  const driver = new MainRenderDriver()
+  const driver = new MainCoreDriver()
   // settings.multithreaded === true
   //   ? new ThreadXRenderDriver({
   //       RendererWorker,
   //     })
-  //   : new MainRenderDriver()
+  //   : new MainCoreDriver()
 
   renderer = new RendererMain(
     {
       appWidth: settings.w || 1920,
       appHeight: settings.h || 1080,
       coreExtensionModule: settings.fontLoader,
+      deviceLogicalPixelRatio:
+        settings.pixelRatio ||
+        screenResolutions[settings.screenResolution] ||
+        screenResolutions[window.innerHeight] ||
+        1,
     },
     target,
     driver