+++ CHANGELOG.md
@@ -1,5 +1,84 @@
 # Changelog
 
+# v0.6.0
+
+_4 jan 2024_
+
+- Added pre-compilation functionality to improve performance. Requires an update to the `vite.config.js` to enable. Read more details in this [blog post](https://lightningjs.io/blogs/blitsPreCompilation.html)
+- Added `defaultFont` to the Launch settings
+
+# v0.5.10
+
+_3 jan 2024_
+
+- Updated and improved documentation
+- Added `screenResolution` and `pixelRatio` options to settings
+
+# v0.5.9
+
+_21 dec 2023_
+
+- Changed transition-end callback to receive real prop value from node
+- Implemented symbol for `wrapper` on Component (freeing up the name wrapper to be used in Component state)
+- Added functionality to prevent a route ending up in the history stack (route option: `inHistory: false`)
+- Improved router backtracking logic
+- Added functionality to override route options during navigation
+- Upgraded to latest version of the renderer (0.6.1) which contains a fix for animations not finishing correctly
+
+# v0.5.8
+
+_15 dec 2023_
+
+- Internal refactor of the `Element`
+
+# v0.5.7
+
+_13 dec 2023_
+
+- Added history and backhandling to the router
+- Linked effects directly to the available shaders exposed by the L3 renderer (starting 0.6.0)
+- Added setting `reactivityMode` to control whether `Proxy` (default) or `defineProperty` is used to trigger reactive side effects
+
+# v0.5.6
+
+_11 dec 2023_
+
+- Upgraded to version 0.6.0 of the Lightning 3 renderer
+
+# v0.5.5
+
+_5 dec 2023_
+
+- Fixed regression in Slots functionality caused by forloop cleanup fix in 0.5.4
+
+# v0.5.4
+
+_5 dec 2023_
+
+- Fixed issue with Components and Elements not always being cleaned up in forloop
+- Added support for nested children inside a forloop on an Element
+
+# v0.5.3
+
+_4 dec 2023_
+
+- Added transition `start` and tranition `end` callbacks
+
+# v0.5.2
+
+_30 nov 2023_
+
+- Fixed issue in code generator leading to creation unnecessary child nodes
+- Added focus handler to router view (that passes focus to the current active page)
+
+# v0.5.1
+
+_29 nov 2023_
+
+- Fixed issue with unexpected unfocus
+- Added fastforward of transitions on the same property
+- Added functionality to skip focus change on key hold
+
 # v0.5.0
 
 _23 nov 2023_